the_internet/requirements.txt
amqp==5.1.1
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
Markdown==3.3.6
celery==5.2.6
requests==2.27.1
django-timezone-field==4.2.3
python-crontab==2.3.4
django-celery-beat # Keep unversioned 